Tests for adding data and fixed logic error
The logic error was due to the functino for creating missing columns
being given __timestamp__ as one of the columns to check for
but since __timestamp__ is not added to the index, the function thinks
it doesn't exist yet and tries to add it. But it does exist in the
actual table, just not in the index.